+++ app/en/resources/integrations/sales/_meta.tsx
@@ -1,18 +1,53 @@
 import type { MetaRecord } from "nextra";
 
 const meta: MetaRecord = {
-  "*": {
-    theme: {
-      breadcrumb: true,
-      toc: true,
-      copyPage: true,
-    },
+  "-- Optimized": {
+    type: "separator",
+    title: "Optimized",
   },
   hubspot: {
     title: "HubSpot",
+    href: "/en/resources/integrations/sales/hubspot",
   },
   salesforce: {
     title: "Salesforce",
+    href: "/en/resources/integrations/sales/salesforce",
+  },
+  "-- Starter": {
+    type: "separator",
+    title: "Starter",
+  },
+  "hubspot-automation-api": {
+    title: "HubSpot Automation API",
+    href: "/en/resources/integrations/sales/hubspot-automation-api",
+  },
+  "hubspot-cms-api": {
+    title: "HubSpot CMS API",
+    href: "/en/resources/integrations/sales/hubspot-cms-api",
+  },
+  "hubspot-conversations-api": {
+    title: "HubSpot Conversations API",
+    href: "/en/resources/integrations/sales/hubspot-conversations-api",
+  },
+  "hubspot-crm-api": {
+    title: "HubSpot CRM API",
+    href: "/en/resources/integrations/sales/hubspot-crm-api",
+  },
+  "hubspot-events-api": {
+    title: "HubSpot Events API",
+    href: "/en/resources/integrations/sales/hubspot-events-api",
+  },
+  "hubspot-marketing-api": {
+    title: "HubSpot Marketing API",
+    href: "/en/resources/integrations/sales/hubspot-marketing-api",
+  },
+  "hubspot-meetings-api": {
+    title: "HubSpot Meetings API",
+    href: "/en/resources/integrations/sales/hubspot-meetings-api",
+  },
+  "hubspot-users-api": {
+    title: "HubSpot Users API",
+    href: "/en/resources/integrations/sales/hubspot-users-api",
   },
 };