+++ toolkit-docs-generator/src/merger/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Merger module exports
+ */
+export * from "./data-merger.js";